requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
DateTime==4.4
discord==2.2.2
discord.py==2.2.2
dnspython==2.2.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jaraco.context==4.1.1
lxml==4.8.0
multidict==6.0.2
pycparser==2.21
pymongo==4.0.2
PyNaCl==1.5.0
pytz==2022.1
requests==2.25.1
rfc3986==1.5.0
reactionmenu==3.1.3
sniffio==1.2.0
soupsieve==2.3.1
typing_extensions==4.1.1
upsidedown==0.4
urllib3==1.26.9
xmltodict==0.12.0
yarl==1.7.2
yt_dlp==2024.11.4
youtube-search==2.1.0
zope.interface==5.4.0
pillow==9.1.1
flask==2.0.2
opuslib==3.0.1
static-ffmpeg==2.2.0
serpapi==0.1.5
google-generativeai==0.8.3